+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - More fine-grained control over the preprocessing, allows for multiple datasets.
 - Angle and grade are now continuous variables.
 - Loss is no longer calculated for the angle and grade.
+- FastAPI now also returns meta-data (prompt, angle, grade, name).
 
 ### Removed